backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
requests==2.32.3
python-dotenv==1.0.1
firebase-admin==6.5.0
google-cloud-firestore==2.17.1
